Added alerts to other deletes and resets

+++ lib/pages/periods/periodForm.dart
@@ -2,6 +2,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:fund_tracker/models/period.dart';
 import 'package:fund_tracker/services/databaseWrapper.dart';
+import 'package:fund_tracker/shared/alert.dart';
 import 'package:fund_tracker/shared/constants.dart';
 import 'package:fund_tracker/shared/library.dart';
 import 'package:fund_tracker/shared/loader.dart';
@@ -42,9 +43,17 @@ class _PeriodFormState extends State<PeriodForm> {
                   textColor: Colors.white,
                   child: Icon(Icons.delete),
                   onPressed: () async {
-                    setState(() => isLoading = true);
-                    DatabaseWrapper(_user.uid).deletePeriod(widget.period);
-                    Navigator.pop(context);
+                    bool hasBeenConfirmed = await showDialog(
+                      context: context,
+                      builder: (BuildContext context) {
+                        return Alert('This custom period will be deleted.');
+                      },
+                    ) ?? false;
+                    if (hasBeenConfirmed) {
+                      setState(() => isLoading = true);
+                      DatabaseWrapper(_user.uid).deletePeriod(widget.period);
+                      Navigator.pop(context);
+                    }
                   },
                 )
               ]
